Move Razorpay initializers to a separate extensions file
- Extracted Razorpay initializers with HTTPClient and Application support from the main Razorpay.swift file
- Created a new Razorpay+Extensions.swift file to organize additional Vapor-related extensions
- Added Application and Request extensions for easy Razorpay client access
- Maintained existing initialization logic for Razorpay client
